Split behavior impl method helpers
